Fix Compiler Warnings
Since most RPi peripherals have 32-bit addresses, C compilers don't like to truncate pointers (which are 64 bits) into 32-bit slots. I added a mechanism using structs to handle this. Unfortunately, I had to use some weird hacks to make it work with arrays (which are not pointers, but "decay" to pointers).

I'm mostly happy, since now I compile with no warnings, but C is a bit obnoxious about not creating an implicit, locally-scoped pointer, hoisted to the top of the scope. That would probably solve loads of weird type errors the C compiler throws our way, but alas.

+++ include/peripherals/dma.h
@@ -64,7 +64,7 @@ typedef struct {
   dma4_channel_regs dma4_regs[4];
 } dma_channel_regs_all;
 
-#define REGS_DMA(channel) ((dma_channel_regs *)((PBASE + 0x7000) + (channel * 0x100)))
+#define REGS_DMA(channel) ((dma_channel_regs *)((u64)((PBASE + 0x7000) + (channel * 0x100))))
 
 #define REGS_DMA_INT_STATUS *((reg32 *)(PBASE + 0x7fe0))
 #define REGS_DMA_ENABLE *((reg32 *)(PBASE + 0x7ff0))
